[refactor] unit test
