+++ src/test/java/racing/RandomMovementTest.java
@@ -1,10 +1,10 @@
 package racing;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racing.domain.RandomMovement;
+import racing.fake.FakeRandom;
 
 import static org.assertj.core.api.Assertions.assertThat;
